packages/substrate-common/pyproject.toml
[project]
name = "substrate-common"
version = "0.1.0"
description = "Shared substrate platform library"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "structlog>=24.4",
    "pyjwt[crypto]>=2.9",
    "httpx>=0.28",
    "python-ulid>=3.0",
    "fastapi>=0.115",
    "starlette>=0.38",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "testcontainers[postgres]>=4.8",
    "ruff>=0.7",
    "mypy>=1.11",
    "vulture>=2.13",
]

[tool.hatch.build.targets.wheel]
packages = ["src/substrate_common"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E","F","W","I","N","UP","B","A","C4","PIE","SIM","RUF","BLE","S110"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
